+++ src/adapters/filter.zig
@@ -1,5 +1,7 @@
 const Iter = @import("../iter.zig").Iter;
 
+const Marker = @import("../meta_extra.zig").Marker;
+
 pub fn Filter(comptime Wrapped: type) type {
     return struct {
         const Self = @This();
@@ -7,6 +9,7 @@ pub fn Filter(comptime Wrapped: type) type {
 
         iter: Iter(Wrapped),
         predicate: *const fn (Item) bool,
+        comptime _: Marker("Filter") = .{},
 
         pub fn next(self: *Self) ?Item {
             while (self.iter.next()) |item| {