+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Restructured code
 - Rename `lazyOnce()` and `lazyRepeat()` to `onceWith()` and `repeatWith()` respectively
 - `assertIsIter()` now returns an error union
+- Use `Marker` for all adapters
+- Rename `assertIsIter()` to `expectImplIter()`
 
 ### Removed